refactor: BonusNumberTest 분리
